removing CommonUtil based on review comments
