+++ samples/src/main/java/org/opensearch/client/samples/PointInTime.java
@@ -12,13 +12,17 @@ import java.util.Collections;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.opensearch.client.opensearch._types.Time;
+import org.opensearch.client.opensearch._types.mapping.IntegerNumberProperty;
+import org.opensearch.client.opensearch._types.mapping.Property;
+import org.opensearch.client.opensearch._types.mapping.TypeMapping;
 import org.opensearch.client.opensearch.core.pit.CreatePitRequest;
 import org.opensearch.client.opensearch.core.pit.CreatePitResponse;
 import org.opensearch.client.opensearch.core.pit.DeletePitRequest;
 import org.opensearch.client.opensearch.core.pit.DeletePitResponse;
 import org.opensearch.client.opensearch.core.pit.ListAllPitResponse;
+import org.opensearch.client.opensearch.indices.CreateIndexRequest;
 import org.opensearch.client.opensearch.indices.DeleteIndexRequest;
-import org.opensearch.client.samples.util.CommonUtil;
+import org.opensearch.client.opensearch.indices.IndexSettings;
 
 /**
  * Run with: <c>./gradlew :samples:run -Dsamples.mainClass=PointInTime</c>
@@ -35,7 +39,19 @@ public class PointInTime {
 
             final var indexName = "my-index";
 
-            CommonUtil.createIndex(client, indexName);
+            if (!client.indices().exists(r -> r.index(indexName)).value()) {
+                LOGGER.info("Creating index {}", indexName);
+                IndexSettings settings = new IndexSettings.Builder().numberOfShards("2").numberOfReplicas("1").build();
+                TypeMapping mapping = new TypeMapping.Builder().properties(
+                        "age",
+                        new Property.Builder().integer(new IntegerNumberProperty.Builder().build()).build()
+                ).build();
+                CreateIndexRequest createIndexRequest = new CreateIndexRequest.Builder().index(indexName)
+                        .settings(settings)
+                        .mappings(mapping)
+                        .build();
+                client.indices().create(createIndexRequest);
+            }
 
             CreatePitRequest createPitRequest = new CreatePitRequest.Builder().targetIndexes(Collections.singletonList(indexName))
                 .keepAlive(new Time.Builder().time("100m").build())
